Fix wrapAssembly extensions not getting rewritten
